+++ eta/core/src/eta/runtime/types/types.h
@@ -2,6 +2,7 @@
 
 #include "symbol.h"
 #include "cons.h"
+#include "atom.h"
 #include "string.h"
 #include "vector.h"
 #include "bytevector.h"